requirements.txt
PyYAML>=5.1.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
ConfigArgParse==1.2.3
configparser==5.0.2
curlify==2.2.1
pytest==5.3.2
pytest-commander==2.4.0
pytest-dependency==0.4.0
pytest-html==3.1.1
pytest-html-reporter==0.2.3
pytest-metadata==1.11.0
pytest-ui==0.3.6b0
pytest-web-ui==1.4.0
requests==2.23.0
selenium==3.141.0
selenium-wire==4.5.2
webdriver-manager==3.4.2
